+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Change history for ui-invoice
 
-## 1.0.0 (IN PROGRESS)
-
+## [1.0.0](https://github.com/folio-org/ui-invoice/tree/v1.0.0) (2019-07-23)
 * New app created with stripes-cli
 * Added default app icon (MODINVOICE-64)
+* [UINV-32](https://issues.folio.org/browse/UINV-32) Invoice approval details set by system
+* [UINV-26](https://issues.folio.org/browse/UINV-26) Search and filter for invoices
+* [UINV-5](https://issues.folio.org/browse/UINV-5) Create and save invoice line details
+* [UINV-7](https://issues.folio.org/browse/UINV-7) Add invoice line based on POL
+* [UINV-3](https://issues.folio.org/browse/UINV-3) Create and save invoice details
+* [UINV-4](https://issues.folio.org/browse/UINV-4) Implement three pane layout for invoice UI